clean up one a
