+++ client/components/EditProduct.jsx
@@ -34,3 +34,5 @@ export default class EditProduct extends Component {
                 )
         }
 }
+
+
