+++ package.json
@@ -35,6 +35,7 @@
         "pg-hstore": "^2.3.2",
         "prop-types": "^15.5.8",
         "react": "^16.0.0",
+        "react-bootstrap": "^0.32.1",
         "react-dom": "^16.0.0",
         "react-redux": "^5.0.2",
         "react-router-dom": "^4.1.1",
